Move paas metrics s3 service into own pkg
