+++ tools/metrics/gauges_s3.go
@@ -1,31 +1,34 @@
 package main
 
 import (
-	"code.cloudfoundry.org/lager"
-	"github.com/aws/aws-sdk-go/service/s3"
 	"time"
+
+	"code.cloudfoundry.org/lager"
+	awss3 "github.com/aws/aws-sdk-go/service/s3"
+
+	"github.com/alphagov/paas-cf/tools/metrics/pkg/s3"
 )
 
 func S3BucketsGauge(
 	logger lager.Logger,
-	s3Service *S3Service,
+	s3Service *s3.S3Service,
 	interval time.Duration,
 ) MetricReadCloser {
 	return NewMetricPoller(interval, func(w MetricWriter) error {
 		metrics := []Metric{}
 
-		buckets, err := s3Service.Client.ListBuckets(&s3.ListBucketsInput{})
+		buckets, err := s3Service.Client.ListBuckets(&awss3.ListBucketsInput{})
 
 		if err != nil {
 			return err
 		}
 
 		metrics = append(metrics, Metric{
-			Kind: Gauge,
-			Time: time.Now(),
-			Name: "aws.s3.buckets.count",
+			Kind:  Gauge,
+			Time:  time.Now(),
+			Name:  "aws.s3.buckets.count",
 			Value: float64(len(buckets.Buckets)),
-			Unit: "count",
+			Unit:  "count",
 		})
 
 		return w.WriteMetrics(metrics)
