+++ tools/metrics/gauges_s3_test.go
@@ -1,37 +1,41 @@
 package main_test
 
 import (
+	"time"
+
 	"code.cloudfoundry.org/lager"
-	. "github.com/alphagov/paas-cf/tools/metrics"
-	"github.com/alphagov/paas-cf/tools/metrics/fakes"
 	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/service/s3"
+	awss3 "github.com/aws/aws-sdk-go/service/s3"
+
+	. "github.com/alphagov/paas-cf/tools/metrics"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	"github.com/onsi/gomega/gbytes"
-	"time"
+
+	"github.com/alphagov/paas-cf/tools/metrics/pkg/s3"
+	"github.com/alphagov/paas-cf/tools/metrics/pkg/s3/fakes"
 )
 
 var _ = Describe("S3 Gauges", func() {
 	var (
-		logger	lager.Logger
-		log		*gbytes.Buffer
-		s3API	*fakes.FakeS3API
-		s3Service *S3Service
+		logger    lager.Logger
+		log       *gbytes.Buffer
+		s3API     *fakes.FakeS3API
+		s3Service *s3.S3Service
 	)
 
-	BeforeEach(func(){
+	BeforeEach(func() {
 		logger = lager.NewLogger("logger")
 		log = gbytes.NewBuffer()
 		logger.RegisterSink(lager.NewWriterSink(log, lager.INFO))
 
 		s3API = &fakes.FakeS3API{}
-		s3Service = &S3Service{Client: s3API}
+		s3Service = &s3.S3Service{Client: s3API}
 	})
 
 	It("returns zero if there are no buckets", func() {
-		s3API.ListBucketsReturns(&s3.ListBucketsOutput{
-			Buckets: []*s3.Bucket{},
+		s3API.ListBucketsReturns(&awss3.ListBucketsOutput{
+			Buckets: []*awss3.Bucket{},
 		}, nil)
 
 		gauge := S3BucketsGauge(logger, s3Service, 1*time.Second)
@@ -50,19 +54,19 @@ var _ = Describe("S3 Gauges", func() {
 		Expect(metric.Kind).To(Equal(Gauge))
 	})
 
-	It("returns the correct number of buckets when there are buckets", func(){
-		s3Buckets := []*s3.Bucket{
-			&s3.Bucket{
+	It("returns the correct number of buckets when there are buckets", func() {
+		s3Buckets := []*awss3.Bucket{
+			&awss3.Bucket{
 				Name: aws.String("bucket-1"),
 			},
-			&s3.Bucket{
+			&awss3.Bucket{
 				Name: aws.String("bucket-2"),
 			},
-			&s3.Bucket{
+			&awss3.Bucket{
 				Name: aws.String("bucket-3"),
 			},
 		}
-		s3API.ListBucketsReturns(&s3.ListBucketsOutput{
+		s3API.ListBucketsReturns(&awss3.ListBucketsOutput{
 			Buckets: s3Buckets,
 		}, nil)
