+++ tools/metrics/pkg/s3/fakes/fake_s3_client.go
